+++ tests/inheritance/test_join_table_inheritance.py
@@ -1,10 +1,9 @@
-from pytest import mark
 import sqlalchemy as sa
 from sqlalchemy_continuum import version_class
 from tests import TestCase
 
 
-class TestJoinTableInheritance(TestCase):
+class JoinTableInheritanceTestCase(TestCase):
     def create_models(self):
         class TextItem(self.Model):
             __tablename__ = 'text_item'
@@ -13,12 +12,15 @@ class TestJoinTableInheritance(TestCase):
             }
             id = sa.Column(sa.Integer, autoincrement=True, primary_key=True)
 
+            name = sa.Column(sa.Unicode(255))
+
             discriminator = sa.Column(
                 sa.Unicode(100)
             )
 
             __mapper_args__ = {
                 'polymorphic_on': discriminator,
+                'with_polymorphic': '*'
             }
 
         class Article(TextItem):
@@ -49,14 +51,16 @@ class TestJoinTableInheritance(TestCase):
         self.ArticleVersion = version_class(self.Article)
         self.BlogPostVersion = version_class(self.BlogPost)
 
-    def test_each_class_has_distinct_version_class(self):
+
+class TestJoinTableInheritance(JoinTableInheritanceTestCase):
+    def test_each_class_has_distinct_version_table(self):
         assert self.TextItemVersion.__table__.name == 'text_item_version'
         assert self.ArticleVersion.__table__.name == 'article_version'
         assert self.BlogPostVersion.__table__.name == 'blog_post_version'
+
         assert issubclass(self.ArticleVersion, self.TextItemVersion)
         assert issubclass(self.BlogPostVersion, self.TextItemVersion)
 
-    @mark.skipif('True')
     def test_each_object_has_distinct_version_class(self):
         article = self.Article()
         blogpost = self.BlogPost()
@@ -75,6 +79,14 @@ class TestJoinTableInheritance(TestCase):
         assert 'transaction_id' in self.TextItemVersion.__table__.c
         assert 'transaction_id' in self.ArticleVersion.__table__.c
         assert 'transaction_id' in self.BlogPostVersion.__table__.c
+
+    def test_with_polymorphic(self):
+        article = self.Article()
+        self.session.add(article)
+        self.session.commit()
+
+        version_obj = self.session.query(self.TextItemVersion).first()
+        assert isinstance(version_obj, self.ArticleVersion)
 
     def test_consecutive_insert_and_delete(self):
         article = self.Article()
@@ -103,3 +115,19 @@ class TestJoinTableInheritance(TestCase):
         assert len(table.primary_key.columns)
         assert 'id' in table.primary_key.columns
         assert 'transaction_id' in table.primary_key.columns
+
+    def test_updates_end_transaction_id_to_all_tables(self):
+        article = self.Article()
+        self.session.add(article)
+        self.session.commit()
+        article.name = u'Updated article'
+        self.session.commit()
+        assert article.versions.count() == 2
+        assert self.session.execute(
+            'SELECT end_transaction_id FROM text_item_version '
+            'ORDER BY transaction_id LIMIT 1'
+        ).scalar()
+        assert self.session.execute(
+            'SELECT end_transaction_id FROM article_version '
+            'ORDER BY transaction_id LIMIT 1'
+        ).scalar()